Editted Activity 2

+++ Tanay_e2.py
@@ -1,12 +1,12 @@
 import sys
 
-if int(sys.argv[1]) >= 220: 
+if float(sys.argv[1]) >= 220: 
     print("Super Typhoon")
-elif int(sys.argv[1]) >= 118 and int(sys.argv[1]) <220:
+elif float(sys.argv[1]) >= 118 and float(sys.argv[1]) <220:
     print("Typhoon")
-elif int(sys.argv[1]) >= 89 and int(sys.argv[1]) <118 :
+elif float(sys.argv[1]) >= 89 and float(sys.argv[1]) <118 :
     print("Severe Tropical Storm")
-elif int(sys.argv[1]) >= 62 and int(sys.argv[1]) <89:
+elif float(sys.argv[1]) >= 62 and float(sys.argv[1]) <89:
     print("Tropical Storm")
-elif int(sys.argv[1]) <=61 :
+elif float(sys.argv[1]) <=61 :
     print("Tropical Depression")
